requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
python-dotenv>=1.0
dnspython>=2.6
requests>=2.32
beautifulsoup4>=4.12
openai>=1.50
apscheduler>=3.10
xhtml2pdf>=0.2.11
